requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Instance Segmentation Models
ultralytics>=8.0.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pycocotools>=2.0.7

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0

# Web Demo
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Google Colab specific (optional)
gdown>=4.7.0
